refactor: extract electron ctx menu
